requirements.txt
# Web App
django==4.0.3
django-extensions==3.1.5
psycopg2==2.9.3
celery==5.2.3

# Django Rest Framework
djangorestframework==3.13.1
markdown==3.3.6
django-filter==21.1
drf-yasg==1.20.0
